% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1ball.R
\name{l1ball}
\alias{l1ball}
\title{l1-ball prior for sparse regression}
\arguments{
\item{y}{A data vector, n by 1}

\item{X}{A design matrix, n by p}

\item{b_w}{The parameter in \eqn{Beta(1, b_w)} for w}

\item{steps}{Number of steps to run the Markov Chain Monte Carlo}

\item{burn-ins}{Number of burn-ins}
}
\value{
trace returns a list of two component: trace_slab and trace_theta, containing all the posterior samples after burn-in.
}
\description{
This package provides function for the l1-ball prior on high-dimensional regression. The main function, l1ball, yields posterior samples for linear regression.
}
\examples{

n = 200
p = 500
X <- matrix(rnorm(n*p),n,p)
d = 5
w0 <- c(rep(0, p-d), rnorm(d)*0.1+1)

y = X%*% w0 + rnorm(n,0,.1)
trace <- l1ball(y,X,b_w = 1.5,2000,burnin = 2000)

plot(colMeans(trace$trace_theta))

}
