% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1ball.R
\name{l1ball}
\alias{l1ball}
\title{l1-ball prior for sparse regression}
\arguments{
\item{y}{A data vector, n by 1}

\item{X}{A design matrix, n by p}

\item{b_w}{The parameter in \eqn{Beta(1, b_w)} for w}

\item{steps}{Number of steps to run the Markov Chain Monte Carlo}

\item{burn-ins}{Number of burn-ins}
}
\value{
trace returns a list of two component: trace_slab and trace_theta, containing all the posterior samples after burn-in.
}
\description{
This package provides function for the l1-ball prior on high-dimensional regression. The main function, l1ball, yields posterior samples for linear regression.
}
\examples{
p=100
X <- diag(p)
d =5
w0 <- c(rep(0, p-d), rnorm(d)+5)
y = X\%*\% w0
trace <- l1ball(y,X,1.5,3000,1000)
}
